chain_of_alpha/requirements.txt
# Chain-of-Alpha MVP Requirements

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Financial data
yfinance>=0.2.0

# Machine learning and LLM
transformers>=4.20.0
torch>=1.12.0
accelerate>=0.20.0

# Backtesting
vectorbt>=0.27.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# API clients (optional)
openai>=1.0.0
requests>=2.28.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# Logging and utilities
python-dotenv>=0.19.0
tqdm>=4.64.0